requirements.txt
rich
pyfiglet
playsound
